Fix class sorting in `capture` liquid tag (#131)
* Refactor liquid script attribute sorting

The Liquid prettier plugin consults the `source` property on various nodes when re-printing the AST as text. Which one gets looked at depends on what type of node is being analyzed and even what ancestor nodes there are. For instance, attributes consult the `source` on the attribute node. But some Liquid Tags end up looking at the source property of the liquid tag instead of the nodes inside it even if they’re modified (for the purposes of attrbute sorting at least).

This new setup allows us to push objects with `source` properties that can all be updated simultanously when sorting the value of an attribute.

* Fix class sorting in `capture` liquid tag

* Update changelog